requirements.txt
Flask
requests
PyJWT
Flask-Limiter
python-dotenv
Flask-WTF
google-auth
google-genai==1.30.0
google-generativeai
gunicorn
pycryptodome
protobuf<5
aiohttp
geopy
gevent